added ctrl+mousewheel to change prop z

+++ src/objects/components/prop.h
@@ -41,7 +41,7 @@ class _Prop {
 		// Attributes
 		const ae::_Texture *Texture;
 		glm::vec4 Color;
-		int Z;
+		float Z;
 		bool Repeat;
 
 	private:
